Business Problem:
The Organisation for Economic Co-operation and Development is creating a department under its name which will focus on how a country can attain sustainable growth. 
This organisation hires you as a data analyst for this department to determine which factors a country should focus on to sustain GDP/ Capita. 
Now, they provide you with the dataset (World Economic Indicator Dataset), which has different countries' and regions' GDPs, populations, and other factors that might impact GDP/Capita. 
Now using Data Analysis methods, develop a solution that can be referred to as “how a country can attain sustainable growth“.